requirements.txt
Django==1.11.15
djangorestframework==3.8.2
gunicorn
django-heroku
Pillow
boto3
whitenoise
django-storages
dj-database-url==0.5.0
djangorestframework-jwt
django-rest-auth[with_social]
django-cors-headers
django-imagekit